Migrated views into own package
