+++ couchdb/couchdb.go
@@ -12,6 +12,7 @@ package couchdb
 //--------------------
 
 import (
+	"encoding/json"
 	"fmt"
 	"reflect"
 	"strings"
@@ -101,9 +102,6 @@ type CouchDB interface {
 	// BulkWriteDocuments allows to create or update many
 	// documents en bloc.
 	BulkWriteDocuments(docs []interface{}, params ...Parameter) (Statuses, error)
-
-	// View performs a view request.
-	View(design, view string, params ...Parameter) ViewResultSet
 }
 
 // couchdb implements CouchDB.
@@ -238,17 +236,20 @@ func (cdb *couchdb) DeleteDatabase(params ...Parameter) ResultSet {
 
 // AllDesigns implements the CouchDB interface.
 func (cdb *couchdb) AllDesigns() ([]string, error) {
-	rs := cdb.Get(cdb.DatabasePath("_all_docs"), nil, StartEndKey("_design/", "_design0/"))
+	jstart, _ := json.Marshal("_design/")
+	jend, _ := json.Marshal("_design0/")
+	startEndKey := Query(KeyValue{"startkey", string(jstart)}, KeyValue{"endkey", string(jend)})
+	rs := cdb.Get(cdb.DatabasePath("_all_docs"), nil, startEndKey)
 	if !rs.IsOK() {
 		return nil, rs.Error()
 	}
-	vr := couchdbViewResult{}
-	err := rs.Document(&vr)
+	designRows := couchdbRows{}
+	err := rs.Document(&designRows)
 	if err != nil {
 		return nil, err
 	}
 	ids := []string{}
-	for _, row := range vr.Rows {
+	for _, row := range designRows.Rows {
 		ids = append(ids, row.ID)
 	}
 	return ids, nil
@@ -265,13 +266,13 @@ func (cdb *couchdb) AllDocuments() ([]string, error) {
 	if !rs.IsOK() {
 		return nil, rs.Error()
 	}
-	vr := couchdbViewResult{}
-	err := rs.Document(&vr)
+	designRows := couchdbRows{}
+	err := rs.Document(&designRows)
 	if err != nil {
 		return nil, err
 	}
 	ids := []string{}
-	for _, row := range vr.Rows {
+	for _, row := range designRows.Rows {
 		ids = append(ids, row.ID)
 	}
 	return ids, nil
@@ -370,12 +371,6 @@ func (cdb *couchdb) BulkWriteDocuments(docs []interface{}, params ...Parameter) 
 		return nil, err
 	}
 	return statuses, nil
-}
-
-// View implements the CouchDB interface.
-func (cdb *couchdb) View(design, view string, params ...Parameter) ViewResultSet {
-	rs := cdb.GetOrPost(cdb.DatabasePath("_design", design, "_view", view), nil, params...)
-	return newViewResultSet(rs)
 }
 
 // idAndRevision retrieves the ID and the revision of the
