+++ views/doc.go
@@ -1,0 +1,12 @@
+// Tideland Go CouchDB Client - Views
+//
+// Copyright (C) 2016-2017 Frank Mueller / Tideland / Oldenburg / Germany
+//
+// All rights reserved. Use of this source code is governed
+// by the new BSD license.
+
+// Package views of the Tideland Go CouchDB Client helps to
+// create and call views.
+package views
+
+// EOF
